[config] Add new context action type override. MER#1541
With this it is possible to define a profile for card which gets
conditionally overriden. For example with following configuration:

[context-rule]
variable     = voicecall_record
value        = equals:enabled
override     = card-name@equals:"mycard", profile:"voicecall", value@constant:"voicecall-record"

The override happens only when mycard active profile is "voicecall" and
voicecall_record context variable changes to value "enabled". When
mycard tries to activate voicecall profile, it is denied. When mycard
tries to activate other profile than voicecall, it is allowed and
override is disabled.