[classify] Have one module per sink or source. Fixes JB#34786
This also helps us avoid unloading/loading same module when changing
routing for both sink and source at the same time. For module unloading,
watch for module unlink events to avoid crashing if extra modules are
loaded when daemon is shutting down. Also if for some reason the module
is unloaded from outside our module we can catch that now.