[classify] Unload modules if route device changes. JB#38988
If target sink or source changes when routing changes, unload the
possible loaded modules from old sink or source. Without the change if
target device changes old devices ports are not updated and the loaded
modules stay lingering. This is a limitation of how the route applying
is implemented. To remedy this, unload modules from old route devices
during routing change if the new route doesn't have identical module
arguments.