[classify] Refactor classify result handling.
Instead of returning a buffer, add the matching types to
new pa_classify_result struct. This way using the elements
elsewhere in the module is simpler.

Also there was a bug in the classify functions' return value,
as the functions would return how many characters were left
in the buf, instead of how many characters long the new string
was.